Refactor: rename DataObjects folders to the Models
