+++ Identity/Hotel.Identity.Infrastructure/Services/DataSeeder.cs
@@ -3,7 +3,7 @@ using Hotel.Identity.Domain.Enums;
 using Hotel.Identity.Domain.Utils;
 using Hotel.Identity.Infrastructure.Abstractions;
 using Hotel.Identity.Infrastructure.Database;
-using Hotel.Identity.Infrastructure.DataObjects;
+using Hotel.Identity.Infrastructure.Models;
 using Microsoft.EntityFrameworkCore;
 
 namespace Hotel.Identity.Infrastructure.Services;
