+++ Reporting/Hotel.Reporting.Domain/Services/ReportService.cs
@@ -2,8 +2,8 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using FluentValidation;
 using Hotel.Reporting.Domain.Abstractions;
-using Hotel.Reporting.Domain.DataObjects;
 using Hotel.Reporting.Domain.Entities;
+using Hotel.Reporting.Domain.Models;
 using Hotel.Shared.Exceptions;
 using Hotel.Shared.MessageContracts;
 using MassTransit;
